fix(bodega): corregir kardex, decimales y origen por item en aprobacion de requisiciones

+++ app/Models/Product.php
@@ -10,7 +10,7 @@ class Product extends Model
     use HasFactory;
 
     protected $casts = [
-        'current_stock' => 'integer',
+        'current_stock' => 'float',
     ];
 
     protected $fillable = [
